app/requirements.txt
requests
psycopg2-binary
datetime
pandas